+++ market/src/database/mongodb/repository/config_exchane.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model } from "mongoose";
 import { Config_exchange, Config_exchangeDocument } from "../schema/config_exchange.schema";
-import { ConfigExchangeDto } from "../../manager/dto/configExchange.dto";
+import { ConfigExchangeDto } from "../../../dto/configExchange.dto";
 
 @Injectable()
 export class Config_exchangeRepository {
