+++ market/src/database/mongodb/schema/status_price.schema.ts
