+++ market/src/database/mongodb/schema/yf_info.schema.ts
