+++ market/src/manager/manager.module.ts
@@ -2,14 +2,14 @@
 import { Module } from '@nestjs/common';
 import { ManagerController } from './manager.controller';
 import { ManagerService } from './manager.service';
-import { YahoofinanceModule } from '../yahoofinance/yahoofinance.module';
-import { MongodbModule } from '../mongodb/mongodb.module';
+import { MarketModule } from '../market/market.module';
+import { MongodbModule } from '../database/mongodb/mongodb.module';
 import { UpdaterModule } from '../updater/updater.module';
 
 @Module({
     imports: [
         UpdaterModule,
-        YahoofinanceModule,
+        MarketModule,
         // HttpModule.register({
         //     timeout: 90000,
         // }),
