+++ market/src/market/py/getInfoByTicker.py
