+++ market/src/market/py/getPriceByTicker.py
