+++ market/src/updater/updater.module.ts
@@ -1,12 +1,12 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 import { UpdaterService } from './updater.service';
-import { YahoofinanceModule } from '../yahoofinance/yahoofinance.module';
-import { MongodbModule } from '../mongodb/mongodb.module';
+import { MarketModule } from '../market/market.module';
+import { MongodbModule } from '../database/mongodb/mongodb.module';
 
 @Module({
   imports: [
-    YahoofinanceModule,
+    MarketModule,
     HttpModule.register({
       timeout: 90000,
     }),
